rename mods tab to moddbupdatestab
