+++ stalker-gamma-gui/App.axaml.cs
@@ -24,7 +24,7 @@ using stalker_gamma.core.Services.GammaInstaller.Utilities;
 using stalker_gamma.core.Utilities;
 using stalker_gamma.core.ViewModels.MainWindow;
 using stalker_gamma.core.ViewModels.Tabs;
-using stalker_gamma.core.ViewModels.Tabs.ModsTab;
+using stalker_gamma.core.ViewModels.Tabs.ModDbUpdatesTab;
 
 namespace stalker_gamma_gui;
 
@@ -71,7 +71,7 @@ public partial class App : Application
                 s.AddScoped<GammaInstaller>();
 
                 s.AddScoped<MainTabVm>();
-                s.AddScoped<ModsTabVm>();
+                s.AddScoped<ModDbUpdatesTabVm>();
                 s.AddScoped<MainWindowVm>();
 
                 var resolver = Locator.CurrentMutable;
