--- stalker-gamma-gui/Controls/Tabs/ModsTab.axaml.cs
@@ -1,16 +1,0 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-using Avalonia.ReactiveUI;
-using stalker_gamma.core.ViewModels.Tabs;
-using stalker_gamma.core.ViewModels.Tabs.ModsTab;
-
-namespace stalker_gamma_gui.Controls.Tabs;
-
-public partial class ModsTab : ReactiveUserControl<ModsTabVm>
-{
-    public ModsTab()
-    {
-        InitializeComponent();
-    }
-}
