+++ stalker-gamma.core/ViewModels/MainWindow/MainWindowVm.cs
@@ -1,13 +1,16 @@
 ﻿using stalker_gamma.core.Services;
 using stalker_gamma.core.ViewModels.Tabs;
-using stalker_gamma.core.ViewModels.Tabs.ModsTab;
+using ModDbUpdatesTabVm = stalker_gamma.core.ViewModels.Tabs.ModDbUpdatesTab.ModDbUpdatesTabVm;
 
 namespace stalker_gamma.core.ViewModels.MainWindow;
 
-public class MainWindowVm(MainTabVm mainTabVm, ModsTabVm modsTabVm, IsBusyService isBusyService)
-    : ViewModelBase
+public class MainWindowVm(
+    MainTabVm mainTabVm,
+    ModDbUpdatesTabVm modDbUpdatesTabVm,
+    IsBusyService isBusyService
+) : ViewModelBase
 {
     public MainTabVm MainTabVm { get; } = mainTabVm;
-    public ModsTabVm ModsTabVm { get; } = modsTabVm;
+    public ModDbUpdatesTabVm ModDbUpdatesTabVm { get; } = modDbUpdatesTabVm;
     public IsBusyService IsBusyService { get; } = isBusyService;
 }
